benchmark/pyproject.toml
[project]
name = "benchmark"
version = "0.1.0"
description = "Benchmark runner for measuring Mountpoint performance while varying configuration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
]
[dependency-groups]
dev = [
    "pytest",
]
